refactor(web): split plugin page data helpers [OPE-377]
